fix: move ask abi method to helpers
